renamed files, updated README.md
