+++ README.md
@@ -1,17 +1,32 @@
-# Obscura
+![Obscura Logo](https://github.com/AxelAlba/Obscura_V2/blob/master/public/img/logo.PNG)
 
-## Authors
+**Obscura** is a social media app for creatives. It's a place to discover ideas, get inspired, and showcase your own work. It's a platform to explore art whether you are a potential client, a fellow creative, or just bored. As Obscura's tagline says, whatever you see.
+
+#### From the community:
+> Finding your next collab partner is better with Obscura. -just another creative
+
+> I found my first hire in Obscura!. -just another client
+
+> Cool stuff bro. -just another bored person
+
+## :joy_cat:Authors
 * Alba, Axel
 * Murillo, Martin
 * Velasco, Dan
 
-## Project setup
+### :floppy_disk:Project setup
 ```
 npm install
 ```
 
-### Start development server
+### :computer:Start development server
 ```
 npm run dev
 ```
 
+### :dart:Features:
+- :heavy_check_mark:Search User by username 
+- :x:Discover Feed (must exclude followings and self)
+- :x:View profiles (needs refactoring once sessions are implemented)
+- :x:Edit own profile settings (needs refactoring once sessions are implemented)
+- :x:Create comment on posts (needs refactoring once sessions are implemented)
