+++ public/js/addComment.js
