+++ views/post.hbs
@@ -1,7 +1,7 @@
 <!-- required css -->
 <link rel="stylesheet" href="/css/posts.css"> <!-- post.hbs css-->
 <!-- required script -->
-<script src="/js/post.js"></script> 
+<script src="/js/addComment.js"></script> 
 <!-- passing data from server to be used in client-side scripting -->
 <input type="hidden" id="pid" value={{post._id}}>
 <!-- Modal -->
